move datamodule.py and example.py into folder datasets, add Readme in dataset folder
