+++ nowcasting_dataset/data_sources/data_source.py
@@ -1,7 +1,7 @@
 from numbers import Number
 import pandas as pd
 import numpy as np
-from nowcasting_dataset.example import Example, to_numpy
+from nowcasting_dataset.dataset.example import Example, to_numpy
 from nowcasting_dataset import square
 import nowcasting_dataset.time as nd_time
 from dataclasses import dataclass, InitVar
@@ -9,7 +9,6 @@ from typing import List, Tuple, Iterable
 import xarray as xr
 import itertools
 import logging
-from typing import Optional
 
 logger = logging.getLogger(__name__)
