+++ tests/test_datamodule.py
@@ -8,12 +8,12 @@ import xarray as xr
 import pytest
 
 import nowcasting_dataset
-from nowcasting_dataset import datamodule
+from nowcasting_dataset.dataset import datamodule
 from nowcasting_dataset.config.load import load_yaml_configuration
-from nowcasting_dataset.datamodule import NowcastingDataModule
-from nowcasting_dataset.example import validate_example
+from nowcasting_dataset.dataset.datamodule import NowcastingDataModule
+from nowcasting_dataset.dataset.example import validate_example
 from nowcasting_dataset.dataset.batch import batch_to_dataset
-from nowcasting_dataset.example import Example
+from nowcasting_dataset.dataset.example import Example
 
 logging.basicConfig(format='%(asctime)s %(levelname)s %(pathname)s %(lineno)d %(message)s')
 _LOG = logging.getLogger("nowcasting_dataset")
